feat: ensure cart product quantity can be decreased

+++ src/actions/decrease-cart-product-quantity/index.ts
@@ -1,16 +1,17 @@
+
 "use server";
 
 import { headers } from "next/headers";
 
 import { auth } from "@/lib/auth";
-import { removeProductFromCartSchema } from "./schema";
+import { decreaseCartProductQuantitySchema } from "./schema";
 import { db } from "@/db";
 import { cartItemTable } from "@/db/schema";
 import { eq } from "drizzle-orm";
 import z from "zod";
 
-export const removeProductFromCart = async (data: z.infer<typeof removeProductFromCartSchema>) => {
-  removeProductFromCartSchema.parse(data);
+export const decreaseCartProductQuantity = async (data: z.infer<typeof decreaseCartProductQuantitySchema>) => {
+  decreaseCartProductQuantitySchema.parse(data);
   const session = await auth.api.getSession({
     headers: await headers(),
   });
@@ -34,5 +35,14 @@ export const removeProductFromCart = async (data: z.infer<typeof removeProductFr
     throw new Error("Unauthorized: Cart does not belong to user");
 
   }
-  await db.delete(cartItemTable).where(eq(cartItemTable.id, cartItem.id))
+
+  if (cartItem.quantity === 1) {
+    await db.delete(cartItemTable).where(eq(cartItemTable.id, cartItem.id))
+    return;
+  }
+
+  await db
+    .update(cartItemTable)
+    .set({ quantity: cartItem.quantity - 1 })
+    .where(eq(cartItemTable.id, cartItem.id));
 }
